Re-roll migrations for life_benefit_detail and life_cap_percentages (spelling)

+++ db/migrate/20170711152701_create_life_benefit_details.rb
@@ -2,11 +2,11 @@ class CreateLifeBenefitDetails < ActiveRecord::Migration
   def change
     create_table :life_benefit_details do |t|
       t.integer :benefit_detail_id
-      t.decimal :subscriber_cap
+      t.integer :subscriber_cap
       t.integer :subscriber_increment
-      t.decimal :spouse_cap
+      t.integer :spouse_cap
       t.integer :spouse_increment
-      t.decimal :dependent_coverage
+      t.integer :dependent_coverage
       t.decimal :dependent_rate
 
       t.timestamps null: false
